Issue/#7/SupportRenameColumn
Added a new class named RanameMap which contains the column mappings from orig to dest
Added a new ColumnInfoMap which contains the old and dest column info.
Added a new Constructor to the Intersection to pass in a list of rename maps.
Updated Insert and Update statements for Orig and Dest tables.